Tests - Break up `test_api` and create one file per test class
