+++ tests/__init__.py
@@ -1,1 +1,5 @@
-""" Test package for snowexsql."""
+from .db_connection import DBConnection
+
+__ALL__ = [
+    "DBConnection"
+]
